+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## v1.1.15 (1.20.1)
 - nerfed impact of Exhaustion
+- added Shielding Armor